Revert "patient updated"
This reverts commit f7d839d974da5f0500b456ab8a6d7a66b9a3068f, reversing
changes made to 27c15f38ecfb0ef09e07d278007737d6195411fd.

+++ HMSApp/src/main/java/com/hms/controller/AppointmentController.java
@@ -2,6 +2,8 @@ package com.hms.controller;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -30,56 +32,56 @@ import jakarta.validation.Valid;
 @RestController
 @RequestMapping("/api/appointments")
 public class AppointmentController {
+	
+	
 
-
-
-	@Autowired
-	private AppointmentService appointmentService;
-
-	@Autowired
-	private DoctorRepository doctorRepository;
-
-	@Autowired
-	private PatientRepository patientRepository;
-
-	@GetMapping("/{apid}")
-	public ResponseEntity<Appointment> getAppointmentById(@PathVariable int apid) throws InvalidEntityException {
-		Appointment appointment = appointmentService.getAppointmentById(apid);
-
-		return new ResponseEntity<>(appointment, HttpStatus.OK);
-	}
-	@PostMapping("/bookAppointment/{doctorId}/{patientId}")
-	public ResponseEntity<?> createAppointment(@PathVariable("doctorId") int doctorId, 
-			@PathVariable("patientId") int patientId, 
-			@Valid @RequestBody Appointment appointment, 
-			BindingResult result) throws InvalidEntityException {
-		if (result.hasErrors()) {
-			StringBuilder errorMessages = new StringBuilder();
-			result.getAllErrors().forEach(error -> errorMessages.append(error.getDefaultMessage()).append("; "));
-			return new ResponseEntity<>(errorMessages.toString(), HttpStatus.BAD_REQUEST);
-		}
-		Doctor doctor = doctorRepository.findById(doctorId).orElseThrow(() -> new InvalidEntityException("Doctor not found"));
-		Patient patient = patientRepository.findById(patientId).orElseThrow(() -> new InvalidEntityException("Patient not found"));
-		appointment.setDoctorObj(doctor);
-		appointment.setPatientObj(patient);
-		Appointment savedAppointment = appointmentService.saveAppointment(appointment);
-		if (savedAppointment != null) {
-			return new ResponseEntity<>(savedAppointment, HttpStatus.CREATED);
-		} else {
-			throw new InvalidEntityException("Error occurred while saving the appointment.");
-		}
-	}
-	@PutMapping("/reschedule/{appointmentId}")
-	public ResponseEntity<String> rescheduleAppointment(
-			@PathVariable int appointmentId,
-			@RequestParam LocalDate newDate,
-			@RequestParam LocalTime newTime) {
-		String response = appointmentService.rescheduleAppointment(appointmentId, newDate, newTime);
-		return ResponseEntity.ok(response);
-	}
-	@DeleteMapping("/cancel/{appointmentId}")
-	public ResponseEntity<String> cancelAppointment(@PathVariable int appointmentId) {
-		String response = appointmentService.cancelAppointment(appointmentId);
-		return ResponseEntity.ok(response);
-	}
+    @Autowired
+    private AppointmentService appointmentService;
+    
+    @Autowired
+    private DoctorRepository doctorRepository;
+    
+    @Autowired
+    private PatientRepository patientRepository;
+    
+    @GetMapping("/{apid}")
+    public ResponseEntity<Appointment> getAppointmentById(@PathVariable int apid) throws InvalidEntityException {
+        Appointment appointment = appointmentService.getAppointmentById(apid);
+        
+        return new ResponseEntity<>(appointment, HttpStatus.OK);
+    }
+    @PostMapping("/bookAppointment/{doctorId}/{patientId}")
+    public ResponseEntity<?> createAppointment(@PathVariable("doctorId") Long doctorId, 
+                                               @PathVariable("patientId") int patientId, 
+                                               @Valid @RequestBody Appointment appointment, 
+                                               BindingResult result) throws InvalidEntityException {
+        if (result.hasErrors()) {
+            StringBuilder errorMessages = new StringBuilder();
+            result.getAllErrors().forEach(error -> errorMessages.append(error.getDefaultMessage()).append("; "));
+            return new ResponseEntity<>(errorMessages.toString(), HttpStatus.BAD_REQUEST);
+        }
+        Doctor doctor = doctorRepository.findById(doctorId).orElseThrow(() -> new InvalidEntityException("Doctor not found"));
+        Patient patient = patientRepository.findById(patientId).orElseThrow(() -> new InvalidEntityException("Patient not found"));
+        appointment.setDoctorObj(doctor);
+        appointment.setPatientObj(patient);
+        Appointment savedAppointment = appointmentService.saveAppointment(appointment);
+        if (savedAppointment != null) {
+            return new ResponseEntity<>(savedAppointment, HttpStatus.CREATED);
+        } else {
+            throw new InvalidEntityException("Error occurred while saving the appointment.");
+        }
+    }
+    @PutMapping("/reschedule/{appointmentId}")
+    public ResponseEntity<String> rescheduleAppointment(
+            @PathVariable int appointmentId,
+            @RequestParam LocalDate newDate,
+            @RequestParam LocalTime newTime) {
+        String response = appointmentService.rescheduleAppointment(appointmentId, newDate, newTime);
+        return ResponseEntity.ok(response);
+    }
+    @DeleteMapping("/cancel/{appointmentId}")
+    public ResponseEntity<String> cancelAppointment(@PathVariable int appointmentId) {
+        String response = appointmentService.cancelAppointment(appointmentId);
+        return ResponseEntity.ok(response);
+    }
 }
